fix(connector): resolve circular import in connector-kit
